requirements.txt
yara-python>=3.11.0